Refactor dialogue IO into several components and helpers
Also make footer icons more consistent (octicon set, 24px)
