+++ public/_sass/parts/buttons.scss
@@ -106,6 +106,10 @@
   }
 }
 
+.btn-icon {
+  margin-right: 1rem;
+}
+
 a.btn-primary {
   color: var(--bright-text-color);
 }
